feat(YSP-770): change ys_localist to use source and id
We will be duplicating the id in localist to both the localist_id and
the event_id fields in an event to move over to using event_id in the
future.  This also auto-creates the "Localist" source if it does not
exist and sets all localist imported events to come from that source.