fix(YSP-797): default anytime view_mode is not present
In a specific cas of a calendar, this view mode does not exist in other
content types in the view tool.  Sometimes this results in nothing being
selected for a view mode, making the default view be a real render of
the page, causing a recursive render of the page if it is included in
the list of items in the view.  This ensures that if you go to a content
type that no longer has the selected view mode, it defaults to a card,
which we will need to make sure is on every content type.  If this ever
changes, we'll need a hash to define that for each content type.