fix(YSP-760): expose save button on inline block media edit
This displays the save button that is hidden by GinLb.  This makes me
wonder the implications or the reason why they chose to do this across
all dialogs, but it does work.  My real worry here is that users might
not realize when they change the image that it will change the image for
that media item--so they're not choosing a new image, they're replacing
the current image across the site.

It also does not update the sidebar since it doesn't refresh the page on
save.