fix(ys_views_basic): disable view caching
There are multiple issues related to the caching of the view block data,
including reusable blocks not excluding the existing node if it was
configured to do so, and the displaying of events that have since
passed.  This is a more heavy-handed approach, allowing a fresh data to
pull from each time.