feat(YSP-938): add content type and CAS identifier to autocomplete
This allows us to use a different search autocomplete for fields like
the reference card entity reference to not only display the title, but
the type of node that it is, as well as an indicator to let users know
it is CAS protected.  This is needed because users will need to be aware
that if they choose a CAS protected page, it will only show when the
user is logged in, which would be rare.